% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlineNewsPopularity.R
\name{generatePredictionsFile}
\alias{generatePredictionsFile}
\title{generatePredictionsFile
Generates predictions on the test data using the model passed as an argument 
and then writes them to a file for submission}
\usage{
generatePredictionsFile(model, data.directory = "")
}
\arguments{
\item{model}{model for generating predictions on test data}

\item{data.directory}{data where online news test data is stored, also where predictions file will be written}
}
\value{
n/a
}
\description{
generatePredictionsFile
Generates predictions on the test data using the model passed as an argument 
and then writes them to a file for submission
}
\examples{
# data.dir <- '~/Projects/kaggle-onlinenewspopularity/data'
# model <- randomForest.train(data.train.full)
# Generates a new predictions file using datetime as file prefix
# generatePredictionsFile(model, data.dir)
}

