% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loglik}
\alias{loglik}
\title{loglik}
\usage{
loglik(model)
}
\arguments{
\item{model}{object which responds to `deviance`}
}
\value{
= -model$deviance/2
}
\description{
loglik
}
\examples{
# loglik(model)

}

