% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{randomForest.train}
\alias{randomForest.train}
\title{randomForest.train}
\usage{
randomForest.train(data.train, label = "popularity")
}
\arguments{
\item{data.train}{matrix or data frame with training data}

\item{label}{column name of the class identified with the row}
}
\value{
randomForest model trained on training data
}
\description{
randomForest.train
}
\examples{
# randomForest.train(data.train)

}

