% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlineNewsPopularity.R
\name{setupOnlineNewsData}
\alias{setupOnlineNewsData}
\title{setupOnlineNewsData
Requires local copies of data from https://inclass.kaggle.com/c/predicting-online-news-popularity/data
Removes id and url from data
Splits data into train and validation sets, randomly}
\usage{
setupOnlineNewsData(directory = "")
}
\arguments{
\item{directory}{local directory with online news datasets (see https://inclass.kaggle.com/c/predicting-online-news-popularity/data)}
}
\value{
list(data.train = data.train, data.validation = data.validation)
}
\description{
setupOnlineNewsData
Requires local copies of data from https://inclass.kaggle.com/c/predicting-online-news-popularity/data
Removes id and url from data
Splits data into train and validation sets, randomly
}
\examples{
data <- setupOnlineNewsData('~/Projects/kaggle-onlinenewspopularity/data')
data.train <- data[['data.train']]
data.validation <- data[['data.validation']]

}

