% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{train.randomForest}
\alias{train.randomForest}
\title{train.randomForest}
\usage{
train.randomForest(data.train, seed = 214, label = "popularity",
  use.classwts = TRUE)
}
\arguments{
\item{data.train}{matrix or data frame with training data}

\item{seed}{if required for recreating exact results}

\item{label}{column name of the class identified with the row}

\item{use.classwts}{boolean for whether or not to use priors on the classes in the forest. Defaults to TRUE}
}
\value{
randomForest model trained on training data
}
\description{
train.randomForest
}
\examples{
# train.randomForest(data.train)

}

