% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlineNewsPopularity.R
\name{training.partitions}
\alias{training.partitions}
\title{training.partitions
Split data into training and validation sets}
\usage{
training.partitions(data.train, p = 0.2)
}
\arguments{
\item{data.train}{a data frame or matrix of training data to split}

\item{p}{portion of data for validation}
}
\value{
data.train and data.validation
}
\description{
training.partitions
Split data into training and validation sets
}
\examples{
train <- remove.extra.cols(news_popularity_training)
test <- news_popularity_test
partitions <- training.paritions(train, p = 0.2)

}

