% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cross.val}
\alias{cross.val}
\title{Runs k-fold cross validation via rolling window.
Any required libraries to run the classification algorithm should already be loaded.}
\usage{
cross.val(model.function, model.args = list(), data.train, no.subsets = 5)
}
\arguments{
\item{model.function}{function name of the classification algorithm call}

\item{model.args}{arguments to pass to the function call}

\item{data.train}{data frame or matrix of training data, depending on what the algorithm requires}

\item{no.subsets}{number of folds to run cross-validation on. Defaults to 5.}
}
\value{
list(rates, mean.rates) success rates and mean success rates of algorithm on each subset
}
\description{
Runs k-fold cross validation via rolling window.
Any required libraries to run the classification algorithm should already be loaded.
}
\examples{
# Example with multinom
}

