% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlineNewsPopularity.R
\name{generate.predictions.file}
\alias{generate.predictions.file}
\title{generatePredictionsFile
Generates predictions on the test data using the model passed as an argument 
and then writes them to a file for submission}
\usage{
generate.predictions.file(model, test.data)
}
\arguments{
\item{model}{model for generating predictions on test data}
}
\value{
filename - filename of generated predictions file (just written to disk)
}
\description{
generatePredictionsFile
Generates predictions on the test data using the model passed as an argument 
and then writes them to a file for submission
}
\examples{
train <- remove.extra.cols(data('news_popularity_training'))
test <- data('news_popularity_test')
model <- train.randomForest(train)
generate.predictions.file(model, test)

}

