% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{randomForest.train}
\alias{randomForest.train}
\title{randomForest.train}
\usage{
\method{randomForest}{train}(data.train, seed = 187, label = "popularity")
}
\arguments{
\item{data.train}{matrix or data frame with training data}

\item{seed}{if required for recreating exact results}

\item{label}{column name of the class identified with the row}
}
\value{
randomForest model trained on training data
}
\description{
randomForest.train
}

