% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{success.rate}
\alias{success.rate}
\title{success.rate}
\usage{
success.rate(predictions, actual)
}
\arguments{
\item{predictions}{vector of predictions}

\item{actual}{vector of actual values}
}
\value{
success rate as float point
}
\description{
success.rate
}
\examples{
# success.rate(model$predicted, data.train$popularity)

}

